# api-contracts/order-service.openapi.yaml
openapi: 3.0.3
info:
  title: Order Service API
  version: 1.0.0
servers:
  - url: http://localhost:9000
paths:
  /api/orders:
    post:
      summary: Create order
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
        '200':
          description: Idempotent replay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
        '409':
          description: Idempotency conflict
    get:
      summary: List orders (current user, or all for admin)
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /api/orders/{id}/cancel:
    patch:
      summary: Cancel order
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cancelled order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Bad request
  /api/orders/outbox/pending:
    get:
      summary: Get pending outbox events (admin)
      tags: [Outbox]
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Pending events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OutboxEvent'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CreateOrderRequest:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
    CreateOrderResponse:
      type: object
      properties:
        idempotentReplay:
          type: boolean
        order:
          $ref: '#/components/schemas/Order'
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
        unitPrice:
          type: number
          format: double
        totalAmount:
          type: number
          format: double
        status:
          type: string
          enum: [CREATED, CANCELLED]
        createdAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        idempotencyKey:
          type: string
          nullable: true
    OutboxEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        aggregateType:
          type: string
        aggregateId:
          type: string
          format: uuid
        eventType:
          type: string
        payload:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
          nullable: true